Refactor function matcher into matchers module and add abstract base class
