+++ sigmatch/__init__.py
@@ -1,3 +1,3 @@
-from sigmatch.function_matcher import FunctionSignatureMatcher as FunctionSignatureMatcher  # noqa: F401
+from sigmatch.matchers.function_matcher import FunctionSignatureMatcher as FunctionSignatureMatcher  # noqa: F401
 from sigmatch.errors import SignatureMismatchError as SignatureMismatchError  # noqa: F401
 from sigmatch.errors import IncorrectArgumentsOrderError as IncorrectArgumentsOrderError  # noqa: F401
