+++ sigmatch/matchers/possible_call_matcher.py
@@ -1,0 +1,5 @@
+from sigmatch.matchers.abstract import AbstractSignatureMatcher
+
+
+class PossibleCallMatcher(AbstractSignatureMatcher):
+    pass
